requirements.txt
asgiref                       3.7.2
black                         23.9.1
certifi                       2023.7.22
charset-normalizer            3.3.0
click                         8.1.7
Django                        4.2.5
django-filter                 23.3
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.0
drf-nested-routers            0.93.4
exceptiongroup                1.1.3
idna                          3.4
iniconfig                     2.0.0
mypy-extensions               1.0.0
packaging                     23.1
pathspec                      0.11.2
Pillow                        10.0.1
pip                           22.0.2
platformdirs                  3.10.0
pluggy                        1.3.0
psycopg2-binary               2.9.7
PyJWT                         2.8.0
pytest                        7.4.2
pytest-django                 4.5.2
pytz                          2023.3.post1
requests                      2.31.0
setuptools                    59.6.0
sqlparse                      0.4.4
tomli                         2.0.1
typing_extensions             4.8.0
urllib3                       2.0.6
